Naturalize request options into the request module
